fix(vroom): redondear matriz a enteros antes de enviarla a VROOM
VROOM exige IsUint() en cada celda de matrices.car (durations/distances/
costs) - un entero no negativo exacto, sin excepciones. Nunca se
redondeaban las duraciones/distancias reales de OSRM (siempre
fraccionarias) antes de escribirlas en la matriz, causando
'Invalid matrix entry' (400) en cualquier lote real con más de un
candidato. Reproducido y confirmado el fix contra la instancia real
de VROOM en Railway.

Defensa en dos capas: redondeo en _build_vroom_request, y
VroomProfileMatrix ahora tipa como list[list[int]] para que Pydantic
rechace localmente cualquier valor fraccionario futuro.

+++ backend/app/services/vroom_service.py
@@ -52,9 +52,16 @@ class VroomVehicle(BaseModel):
 
 
 class VroomProfileMatrix(BaseModel):
-    durations: list[list[float]]
-    distances: list[list[float]]
-    costs: list[list[float]]
+    # VROOM (rapidjson) exige IsUint() por celda -- entero no negativo exacto
+    # -- y responde 400 {"code":2,"error":"Invalid matrix entry."} ante
+    # cualquier valor con parte fraccionaria (input_parser.cpp:get_matrix en
+    # VROOM-Project/vroom). int en vez de float aqui hace que un valor sin
+    # redondear (p.ej. una duracion real de OSRM como 312.7) falle en la
+    # construccion de este modelo, localmente y con un error claro, en vez de
+    # descubrirse recien en la respuesta HTTP del proveedor.
+    durations: list[list[int]]
+    distances: list[list[int]]
+    costs: list[list[int]]
 
 
 class VroomOptimizationRequest(BaseModel):
